chore: fix dependency imports and refactor
