+++ src/bridging/types.ts
@@ -1,9 +1,9 @@
 import { latest as latestAppData } from '@cowprotocol/app-data'
 import { ethers } from 'ethers'
 import { ChainInfo, ChainId } from '../common'
-import { TokenInfo } from '../common/tokens'
+import { TokenInfo } from '../common/types/tokens'
 import { Address, OrderKind } from '../order-book'
-import { EvmCall } from '../common/transaction'
+import { EvmCall } from '../common/types/ethereum'
 
 export interface BridgeProviderInfo {
   name: string
