+++ src/common/index.ts
@@ -1,7 +1,13 @@
-export * from './chains'
-export * from './configs'
-export * from './cow-error'
-export * from './ipfs'
-export * from './consts'
-export * from './transaction'
-export * from './tokens'
+// Types
+export * from './types/chains'
+export * from './types/config'
+export * from './types/cow-error'
+export * from './types/ethereum'
+export * from './types/tokens'
+
+// Constants
+export * from './consts/chains'
+export * from './consts/config'
+export * from './consts/contracts'
+export * from './consts/ipfs'
+export * from './consts/tokens'
