+++ src/common/utils/config.ts
@@ -1,0 +1,18 @@
+import { ALL_SUPPORTED_CHAIN_IDS } from '../consts/chains'
+import { SupportedChainId } from '../types/chains'
+
+export function mapSupportedNetworks<T>(value: (chainId: SupportedChainId) => T): Record<SupportedChainId, T>
+export function mapSupportedNetworks<T>(value: T): Record<SupportedChainId, T>
+export function mapSupportedNetworks<T>(value: T | ((chainId: SupportedChainId) => T)): Record<SupportedChainId, T> {
+  return ALL_SUPPORTED_CHAIN_IDS.reduce<Record<number, T>>(
+    (acc, chainId) => ({
+      ...acc,
+      [chainId]: typeof value === 'function' ? (value as (chainId: SupportedChainId) => T)(chainId) : value,
+    }),
+    {}
+  )
+}
+
+export function mapAddressToSupportedNetworks(address: string): Record<SupportedChainId, string> {
+  return mapSupportedNetworks(address)
+}
