+++ src/cow-shed/CoWShedHooks.spec.ts
@@ -2,7 +2,7 @@ import { Wallet } from 'ethers'
 import { CowShedHooks } from './CoWShedHooks'
 import { AddressZero } from '@ethersproject/constants'
 import { ICoWShedCall } from './types'
-import { COW_SHED_FACTORY, COW_SHED_IMPLEMENTATION } from '../common/consts'
+import { COW_SHED_FACTORY, COW_SHED_IMPLEMENTATION } from '../common'
 import { COW_SHED_PROXY_INIT_CODE } from './proxyInitCode'
 import { SigningScheme } from '@cowprotocol/contracts'
 import { formatBytes32String, solidityPack } from 'ethers/lib/utils'
