+++ src/order-book/api.spec.ts
@@ -11,11 +11,11 @@ jest.mock('cross-fetch', () => {
 })
 
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock'
-import { CowError } from '../common/cow-error'
+import { CowError } from '../common/types/cow-error'
 import { OrderBookApi } from './api'
 import { BuyTokenDestination, EcdsaSigningScheme, OrderKind, SellTokenSource, SigningScheme } from './generated'
-import { SupportedChainId } from '../common/chains'
-import { ETH_ADDRESS } from '../common/consts'
+import { SupportedChainId } from '../common/types/chains'
+import { ETH_ADDRESS } from '../common/consts/contracts'
 import { AUCTION } from './mock'
 
 enableFetchMocks()
