+++ src/order-signing/utils.ts
@@ -21,7 +21,7 @@ import {
 import type { Signer } from '@ethersproject/abstract-signer'
 import type { SigningResult, SignOrderParams, SignOrderCancellationParams, UnsignedOrder } from './types'
 
-import { COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS } from '../common/consts'
+import { COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS } from '../common/consts/contracts'
 import { CowError, SupportedChainId } from '../common'
 import { EcdsaSigningScheme } from '../order-book'
 import { SignOrderCancellationsParams } from './types'
@@ -249,6 +249,12 @@ export async function generateOrderId(
   const domain = await getDomain(chainId)
   const orderDigest = hashOrder(domain, order)
   // Generate the orderId from owner, orderDigest, and max validTo
+  console.log('params GEN', {
+    ...params,
+    orderDigest,
+    // Different validTo when signing because EthFlow contract expects it to be max for all orders
+    validTo: order.validTo,
+  })
   const orderId = packOrderUidParams({
     ...params,
     orderDigest,
